Add .gitignore for Visual Studio and JetBrains IDEs
This commit adds a comprehensive .gitignore file that covers temporary files, build results, user-specific files, and other files which should not be checked into source control, that are created by both Visual Studio and JetBrains IDEs. The .gitignore file is sourced from the one maintained by GitHub at https://github.com/github/gitignore, and it is important to reduce the noise in the project's change history and make it easier for contributors to participate in the project. It also helps prevent the accidental commit of sensitive information like passwords or keys.