Update ROLAbsX() method in Cpu.Rotate.cs
Updated ROLAbsX() method in Cpu.Rotate.cs to ensure correct address is calculated. Previously, the method was only considering the next 16-bit program for calculation. The method has been modified to add 'X' to the output of 'ReadNext16BitProgram()' to retrieve the correct 'address'. The change leads to correct read address and avoids potential issues in CPU instructions.